requirements.txt
cryptography==42.0.5
pyjwt[crypto]==2.8.0
python-dotenv==1.0.1
httpx==0.27.0
argparse==1.4.0
pytest==8.2.2
requests==2.32.3
fastapi
uvicorn[standard]
sqlalchemy
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
langchain-openai
langchain-community
langchain-google-genai
python-dotenv
hushh_mcp
faiss-cpu
requests
werkzeug